+++ Bundles/LeafVisu/visuVTKAdaptor/src/visuVTKAdaptor/STransform.cpp
@@ -140,7 +140,7 @@ void STransform::updateFromVtk()
         {
             for(int ct = 0; ct < 4; ct++)
             {
-                trf->setCoefficient(lt, ct, mat->GetElement(lt, ct));
+                trf->setCoefficient(static_cast<size_t>(lt), static_cast<size_t>(ct), mat->GetElement(lt, ct));
             }
         }
     }
@@ -177,7 +177,7 @@ void STransform::updating()
         {
             for(int ct = 0; ct < 4; ct++)
             {
-                mat->SetElement(lt, ct, trf->getCoefficient(lt, ct));
+                mat->SetElement(lt, ct, trf->getCoefficient(static_cast<size_t>(lt), static_cast<size_t>(ct)));
             }
         }
     }
